Ignore additional directories
The directories temp, out and styleguide do not need to be tracked by Git. styleguide is another output directory.  There is a directory info-model/rules/temp which would also be excluded by the changed .gitignore rules. I do not understand for what the directory is needed and whether it is a "real" temp directory. It is NOT created when publishing all the defined transformation scenarios.